feat(host): add 128-bit trace identities and resume lineage
Replace narrow trace identifiers with randomly generated 128-bit identities and serialize them as fixed-width canonical hexadecimal values across host policy and stream boundaries.

Record an optional parent trace for resumed executions, provide explicit resumed-context construction, and update host security and live integration coverage for the wider identity type.

+++ crates/etas_host/src/stream/store.rs
@@ -1293,7 +1293,7 @@ mod tests {
             id: HostRequestId(id),
             operation,
             authority: AuthorityContext::deny_all(),
-            trace: TraceContext::root(TraceId(id)),
+            trace: TraceContext::root(TraceId(u128::from(id))),
             budget: ExecutionBudget::default(),
         }
     }
